pyproject.toml
[project]
name = "nlba"
version = "0.1.0"
description = "Natural Language Bash Assistant"
authors = [{ name = "Artyomka", email = "artyomka@example.com" }]
dependencies = [
    "google-generativeai",
    "openai"
]

requires-python = ">=3.9"

[project.optional-dependencies]
dev = ["pytest"]

[build-system]
requires = ["uv>=0.1.0", "setuptools"]
build-backend = "setuptools.build_meta"


